backend/requirements-rhino.lock.txt
# Generated from pyenv "rhino": ~/.pyenv/versions/rhino/bin/pip freeze | LC_ALL=C sort -u
# Rebuild: pyenv shell rhino && pip freeze | LC_ALL=C sort -u > requirements-rhino.lock.txt
#
Brotli==1.1.0
Cython==3.1.3
GitPython==3.1.45
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
PyOpenGL==3.1.9
PyWavelets==1.9.0
PyYAML==6.0.2
Pygments==2.19.2
PytorchWildlife==1.2.4
SQLAlchemy==2.0.36
Werkzeug==3.1.3
absl-py==2.3.1
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asttokens==3.0.0
astunparse==1.6.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==5.0.0
boto3==1.40.21
botocore==1.40.21
certifi==2025.8.3
cffi==2.0.0
chainer==7.8.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.35.2
distro==1.9.0
docutils==0.22
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.6
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.12
google-pasta==0.2.0
gradio==5.44.1
gradio_client==1.12.1
greenlet==3.3.2
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.37.0
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
keras==3.11.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
lightning==2.5.4
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
matplotlib==3.10.5
mdurl==0.1.2
meshzoo==0.12.0
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
munch==4.0.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
openai==1.55.3
opencv-python-headless==4.10.0.84
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pandas==2.3.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pi_heif==1.1.0
pillow-avif-plugin==1.5.2
pillow==11.0.0
polars==1.33.0
pretrainedmodels==0.7.4
progressbar2==4.5.0
progressbar==2.5
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.2
pybboxes==0.1.6
pycparser==2.21
pydantic-settings==2.6.1
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.18
python-utils==3.9.1
pytorch-lightning==2.5.4
pytz==2025.2
regex==2025.11.3
requests-toolbelt==1.0.0
requests==2.32.4
rich==14.1.0
roboflow==1.2.7
rsa==4.9.1
ruff==0.12.11
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
sahi==0.11.34
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
statistics==1.0.3.5
stonefish-runtime==0.3.3
supervision==0.23.0
sympy==1.14.0
tensorboard-data-server==0.7.2
tensorboard==2.20.0
tensorflow==2.20.0
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.19
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.1
torchnet==0.0.4
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics-thop==2.0.16
ultralytics==8.3.202
urllib3==2.5.0
uvicorn==0.32.1
uvloop==0.22.1
visdom==0.2.4
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wget==3.2
wheel==0.45.1
wrapt==1.17.3
yacs==0.1.8
yarl==1.20.1
yolov5==7.0.13
zipp==3.23.0
